+++ src/TrackYourDay.MAUI/ServiceRegistration/ServiceCollections.cs
@@ -1,12 +1,15 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
+using Quartz;
 using TrackYourDay.Core;
 using TrackYourDay.Core.Activities;
 using TrackYourDay.Core.Activities.ActivityRecognizing;
 using TrackYourDay.Core.Breaks;
 using TrackYourDay.Core.Settings;
 using TrackYourDay.Core.Versioning;
+using TrackYourDay.Core.Workdays;
 using TrackYourDay.MAUI.BackgroundJobs;
+using TrackYourDay.MAUI.BackgroundJobs.ActivityTracking;
 using TrackYourDay.MAUI.Notifications;
 
 namespace TrackYourDay.MAUI.ServiceRegistration
@@ -17,7 +20,8 @@ namespace TrackYourDay.MAUI.ServiceRegistration
         {
             services.AddScoped<ISystemStateRecognizingStrategy, DefaultActivityRecognizingStategy>();
             // Refactor to avoid this in future
-            services.AddSingleton<ActivityTracker>(container => {
+            services.AddSingleton<ActivityTracker>(container =>
+            {
                 var clock = container.GetRequiredService<IClock>();
                 var publisher = container.GetRequiredService<IPublisher>();
                 var startedActivityRecognizingStrategy = new DefaultActivityRecognizingStategy();
@@ -41,6 +45,7 @@ namespace TrackYourDay.MAUI.ServiceRegistration
 
         public static IServiceCollection AddNotifications(this IServiceCollection services)
         {
+            services.AddSingleton<WorkdayReadModelRepository>();
             services.AddSingleton<ExecutableNotificationFactory>();
             services.AddSingleton<NotificationRepository>();
             services.AddSingleton<NotificationService>();
@@ -58,6 +63,26 @@ namespace TrackYourDay.MAUI.ServiceRegistration
                 serviceProvider.GetService<SettingsService>().GetCurrentSettingSet());
 
             return services;
+        }
+
+        public static IServiceCollection AddBackgroundJobs(this IServiceCollection services)
+        {
+            return services.AddQuartz(q =>
+            {
+                q.UseMicrosoftDependencyInjectionJobFactory();
+
+                q.ScheduleJob<ActivityEventTrackerJob>(trigger => trigger
+                    .WithIdentity("Activity Recognizing Job")
+                    .WithDescription("Job that periodically recognizes user activities")
+                    .WithDailyTimeIntervalSchedule(x => x.WithInterval((int)ActivitiesSettings.CreateDefaultSettings().FrequencyOfActivityDiscovering.TotalSeconds, IntervalUnit.Second))
+                    .StartNow());
+
+                q.ScheduleJob<NotificationsProcessorJob>(trigger => trigger
+                    .WithIdentity("Notifications Processor Job")
+                    .WithDescription("Job that periodically checks for notifications to process")
+                    .WithDailyTimeIntervalSchedule(x => x.WithInterval(10, IntervalUnit.Second))
+                    .StartNow());
+            });
         }
     }
 }
